openface/requirements-python2.txt
Flask == 0.12.4
python-crontab == 2.3.6
scikit-learn >= 0.17, < 0.18
numpy >= 1.1, < 2.0
imutils